chore: move parser functions to new location and clean up code
